+++ src/simtools/production_configuration/production_grid_job_rows.py
@@ -1,5 +1,5 @@
 """
-Builds backend-agnostic job row dictionaries for simulation production grids.
+Build simulation execution rows based on the production configuration and lookup tables.
 
 A job row is a configuration dictionary for a single simulation job/run, containing all
 parameters needed to launch that simulation (e.g., primary, zenith, azimuth, energy range,
@@ -8,6 +8,7 @@ model version, etc.).
 
 import itertools
 
+import numpy as np
 from astropy import units as u
 
 from simtools.production_configuration.corsika_limits_lookup import CorsikaLimitsLookup
@@ -67,25 +68,70 @@ def get_core_scatter_max_for_zenith_angle(
     return min(configured_scatter_max, lookup_scatter_max.to(configured_scatter_max.unit))
 
 
-def build_backend_agnostic_job_rows(
-    args_dict,
-    calculate_scaled_nshow,
-    get_energy_range_for_zenith_angle_function=get_energy_range_for_zenith_angle,
-    get_core_scatter_max_for_zenith_angle_function=get_core_scatter_max_for_zenith_angle,
+def calculate_log_energy_midpoint(energy_range_pair):
+    """Return the geometric-mean energy for an energy range pair."""
+    energy_min, energy_max = energy_range_pair
+
+    if not isinstance(energy_min, u.Quantity) or not isinstance(energy_max, u.Quantity):
+        raise TypeError("energy_range_pair must contain astropy Quantity values.")
+
+    energy_min_tev = energy_min.to(u.TeV)
+    energy_max_tev = energy_max.to(u.TeV)
+
+    if energy_min_tev <= 0 * u.TeV or energy_max_tev <= 0 * u.TeV:
+        raise ValueError("Energy range values must be strictly positive.")
+
+    mean_log_energy = np.mean(
+        [
+            np.log10(energy_min_tev.value),
+            np.log10(energy_max_tev.value),
+        ]
+    )
+    return 10**mean_log_energy * u.TeV
+
+
+def calculate_scaled_nshow(
+    energy_range_pair,
+    baseline_nshow,
+    nshow_power_index=None,
+    reference_energy=None,
 ):
+    """Return an energy-dependent nshow value."""
+    if baseline_nshow < 1:
+        raise ValueError("baseline_nshow must be a positive integer.")
+
+    if nshow_power_index is None:
+        return baseline_nshow
+
+    if reference_energy is None:
+        raise ValueError("reference_energy is required when nshow_power_index is configured.")
+
+    midpoint_energy = calculate_log_energy_midpoint(energy_range_pair)
+    scaling_factor = (midpoint_energy / reference_energy.to(midpoint_energy.unit)).to_value(
+        u.dimensionless_unscaled
+    ) ** nshow_power_index
+    scaled_nshow = int(np.ceil(baseline_nshow * scaling_factor))
+
+    if scaled_nshow < 1:
+        raise ValueError("Scaled nshow must be at least 1.")
+
+    return scaled_nshow
+
+
+def build_simulation_execution_rows(args_dict):
     """
-    Build normalized production-grid rows for backend consumers.
+    Build simulation execution rows based on the production configuration and lookup tables.
+
+    Each row corresponds to a single simulation job/run and contains all parameters needed to
+    launch that simulation (e.g., primary, zenith, azimuth, energy range, model version, etc.).
+    The energy range and core-scatter radius can be adjusted based on the direction-dependent
+    lookup-table limits. The number of events (``nshow``) can be scaled with energy according
+    to a configurable power law.
 
     Parameters
     ----------
     args_dict : dict
         Production-job configuration.
-    calculate_scaled_nshow : callable
-        Callback used to compute the per-row ``nshow`` value.
-    get_energy_range_for_zenith_angle_function : callable, optional
-        Callback used to derive the direction-dependent energy range.
-    get_core_scatter_max_for_zenith_angle_function : callable, optional
-        Callback used to derive the direction-dependent core-scatter radius.
 
     Returns
     -------
@@ -135,7 +181,7 @@ def build_backend_agnostic_job_rows(
         corsika_he,
         energy_range_pair,
     ) in combinations:
-        selected_energy_range_pair = get_energy_range_for_zenith_angle_function(
+        selected_energy_range_pair = get_energy_range_for_zenith_angle(
             zenith,
             energy_range_pair,
             corsika_limits,
@@ -144,7 +190,7 @@ def build_backend_agnostic_job_rows(
         if selected_energy_range_pair is None:
             continue
 
-        selected_core_scatter_max = get_core_scatter_max_for_zenith_angle_function(
+        selected_core_scatter_max = get_core_scatter_max_for_zenith_angle(
             zenith,
             core_scatter,
             corsika_limits,
@@ -154,7 +200,7 @@ def build_backend_agnostic_job_rows(
             selected_energy_range_pair, nshow, nshow_power_index, reference_energy
         )
 
-        for _ in range(number_of_runs):
+        for row_index in range(number_of_runs):
             rows.append(
                 {
                     "primary": primary,
@@ -171,5 +217,4 @@ def build_backend_agnostic_job_rows(
                     "run_number": run_number + row_index,
                 }
             )
-            row_index += 1
     return rows
